v0.86.c: C++ API and CLI/C++ Interface
improved windows build guide, fixed a bug in the .net wrapper

fixed some bugs in the windows build files

work in getting c++ api

finshed c++ api wrapper

hopefully mostly working

half-way making OTAPI_Basic

more work... will squash later

finished OTAPI_Basic

deleted old OTAPI_funcdef.h and replaced with OTAPI_Basic, updated swig

now working with moneychanger

more cleanup of code

cleaned up code, re-coded CLI (.net) api

more work for developing cli/c++ interface

fixed makefile

ot cli dll now fully working!

+++ src/otlib/OTASCIIArmor.cpp
@@ -1125,7 +1125,7 @@ bool OTASCIIArmor::SetAndPackString(const OTString & strData, bool bLineBreaks) 
 //                                   uint32_t src_length,
 //                                   bool     bZeroSource/*=false*/) // if true, sets the source buffer to zero after copying is done.
 
-    OTPassword::safe_memcpy(pSource, lSourcelen, pUint, theSize);
+    OTPassword::safe_memcpy(pSource, lSourcelen, pUint, static_cast<uint32_t>(theSize));
 //	memcpy(pSource, static_cast<const unsigned char*>(pUint), theSize );
 	
 	// Now we are compressing first before base64-encoding (for strings, anyway)
